refactor: 💡 move plugins to their own folder
